+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 2.8.5
+### Unreleased
 * fix http not implemented uri with query args 
 * bootstrap: fix error wrapping from govalidator when schema of remote config has required map field
 ### v2.8.4